[jacodb-storage, jacodb-core] Move pluggable caches to jacodb-storage
Pluggable caches API moved to the jacodb-storage-api module, implementations
are in the jacodb-storage module. The jacodb-storage module has compileOnly
dependencies on the libraries providing caches. So to use caches without
dependency on the jacodb-core module, one should define particular
dependency of the caching library, Guava or Xodus.
